+++ src/Recyclarr.Cli/Console/Settings/IListCustomFormatSettings.cs
@@ -1,0 +1,10 @@
+using Recyclarr.TrashLib.Config;
+
+namespace Recyclarr.Cli.Console.Settings;
+
+public interface IListCustomFormatSettings
+{
+    SupportedServices Service { get; }
+    bool ScoreSets { get; }
+    bool Raw { get; }
+}